requirements.txt
Flask
pandas
PyYAML
python-docx
PyPDF2
scikit-learn
torch
transformers
tqdm
gunicorn
protobuf
streamlit
sentence-transformers
